Refactor of render pipeline to prevent some redundencies in element creations

+++ Engine/Code/Engine/Renderer/RenderBuffer.cpp
@@ -100,7 +100,7 @@ bool RenderBuffer::Create( size_t dataByteSize, size_t elementByteSize )
 	ID3D11Device* device = m_owner->m_device;
 
 	D3D11_BUFFER_DESC desc;
-	desc.ByteWidth = dataByteSize;
+	desc.ByteWidth = static_cast<UINT>( dataByteSize );
 	desc.Usage = ToDXMemoryUsage( m_memHint );
 	desc.BindFlags = ToDXUsage( m_usage );
 	
@@ -115,7 +115,7 @@ bool RenderBuffer::Create( size_t dataByteSize, size_t elementByteSize )
 	}
 
 	desc.MiscFlags = 0; //Used in special cases
-	desc.StructureByteStride = elementByteSize;
+	desc.StructureByteStride = static_cast<UINT>( elementByteSize );
 
 	device->CreateBuffer( &desc, nullptr, &m_handle );
 	return m_handle != nullptr;
